+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a configuration file example to the README.
+- Improve the project's description in the README.
 
 ### Fixed