+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.4] - 2023-10-29
 
 ### Added
 
+- Implement the "inspect-repo" command that checks an entire repository.
 - Add a configuration file example to the README.
 - Improve the project's description in the README.
 - Add installation instructions to the README.